+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `FormulaContext.onEvent` to replace `FormulaContext.callback` and `FormulaContext.eventCallback`
 - Add Kotlin coroutine support! New `formula-coroutines` module to provide `toFlow()` extension function and `FlowStream`.
 - Add `Listener` type to consolidate event handling types `() -> Unit` and `(Event) -> Unit` 
+- Rename and move `FormulaContext.UpdateBuilder` into `StreamBuilder`. 
 
 ## [0.7.0] - June 30, 2021
 - **Breaking**: Remove `events(observable) { }` utility function.